requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
 
streamlit==1.54.0
streamlit-drawable-canvas-fix==0.9.8
 
torch==2.1.2+cpu
torchvision==0.16.2+cpu
 
paddleocr==2.7.0.3
paddlepaddle==2.6.2
 
opencv-python-headless==4.6.0.66
numpy==1.26.4
pandas==2.3.3
pillow==10.4.0
openpyxl==3.1.5
requests==2.32.5